+++ apps/cow-fi/types/addToWalletState.ts
@@ -1,0 +1,7 @@
+export type AddToWalletStateValues = 'unknown' | 'adding' | 'added' | 'error' | 'takingTooLong' | 'connecting'
+
+export interface AddToWalletState {
+  state: AddToWalletStateValues
+  errorMessage?: string
+  autoConnect: boolean
+}
